Consolidate noise UGens
