--- supriya/ugens/Crackle.py
@@ -1,27 +1,0 @@
-import collections
-
-from supriya import CalculationRate
-from supriya.synthdefs import UGen
-
-
-class Crackle(UGen):
-    """
-    A chaotic noise generator.
-
-    ::
-
-        >>> crackle = supriya.ugens.Crackle.ar(
-        ...     chaos_parameter=1.25,
-        ...     )
-        >>> crackle
-        Crackle.ar()
-
-    """
-
-    ### CLASS VARIABLES ###
-
-    __documentation_section__ = "Noise UGens"
-
-    _ordered_input_names = collections.OrderedDict([("chaos_parameter", 1.5)])
-
-    _valid_calculation_rates = (CalculationRate.AUDIO, CalculationRate.CONTROL)
